% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaDfScaleFactor.R
\name{ConstScaleFactor}
\alias{ConstScaleFactor}
\title{Constant scale factor for differential evolution.}
\usage{
ConstScaleFactor(lF)
}
\arguments{
\item{lF}{Local configuration.}
}
\value{
A constant scale factor.
}
\description{
Constant scale factor for differential evolution.
}
\examples{
parm<-function(x){function() {return(x)}}
lF<-list()
lF$ScaleFactor1<-parm(0.90)
ConstScaleFactor(lF)
ConstScaleFactor(lF)
}
\seealso{
Other ScaleFactor: 
\code{\link{UniformRandomScaleFactor}()}
}
\concept{ScaleFactor}
