% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaDfScaleFactor.R
\name{UniformRandomScaleFactor}
\alias{UniformRandomScaleFactor}
\title{Uniform random scale factor for differential evolution.}
\usage{
UniformRandomScaleFactor(lF)
}
\arguments{
\item{lF}{Local configuration.}
}
\value{
A constant scale factor.
}
\description{
The scale factor is drawn from 
             \code{0.000001} to \code{1.0}.
}
\examples{
parm<-function(x){function() {return(x)}}
lF<-list()
UniformRandomScaleFactor(lF)
UniformRandomScaleFactor(lF)
}
\seealso{
Other Scale Factor: 
\code{\link{CauchySF}()},
\code{\link{ConstScaleFactor}()},
\code{\link{DETVSF}()},
\code{\link{FitnessBasedSelfAdaptiveSF}()},
\code{\link{RandomGaussianSF}()},
\code{\link{UniformRandomScaleFactorDERSF}()}
}
\concept{Scale Factor}
