% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaDfDecode.R
\name{xegaDfDecodeGene}
\alias{xegaDfDecodeGene}
\title{Decode a gene}
\usage{
xegaDfDecodeGene(gene, lF)
}
\arguments{
\item{gene}{Real gene.}

\item{lF}{Local configuration of the genetic algorithm.}
}
\value{
Decoded gene.
}
\description{
\code{xegaDfDecodeGene()} decodes a real gene.
}
\details{
\code{xegaDfDecodeGene()} is the identity function.
}
\examples{
gene<-xegaDfInitGene(lFxegaDfGene)
xegaDfDecodeGene(gene, lFxegaDfGene)

}
\seealso{
Other Decoder: 
\code{\link{xegaDfGeneMapIdentity}()}
}
\concept{Decoder}
