% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaDfScaleFactor.R
\name{ConstScaleFactor}
\alias{ConstScaleFactor}
\title{Constant scale factor for differential evolution.}
\usage{
ConstScaleFactor(lF)
}
\arguments{
\item{lF}{Local configuration.}
}
\value{
A constant scale factor.
}
\description{
Constant scale factor for differential evolution.
}
\examples{
parm<-function(x){function() {return(x)}}
lF<-list()
lF$ScaleFactor1<-parm(0.90)
ConstScaleFactor(lF)
lF$ScaleFactor1<-parm(1.10)
ConstScaleFactor(lF)
}
\seealso{
Other Scale Factor: 
\code{\link{CauchySF}()},
\code{\link{DETVSF}()},
\code{\link{FitnessBasedSelfAdaptiveSF}()},
\code{\link{RandomGaussianSF}()},
\code{\link{UniformRandomScaleFactor}()},
\code{\link{UniformRandomScaleFactorDERSF}()}
}
\concept{Scale Factor}
