% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaDfInitGene.R
\name{xegaDfInitGene}
\alias{xegaDfInitGene}
\title{Initialize a real-coded gene.}
\usage{
xegaDfInitGene(lF)
}
\arguments{
\item{lF}{Local configuration of the genetic algorithm.}
}
\value{
A real-coded gene (a named list):
        \itemize{
        \item \code{$evaluated}: FALSE. See package \code{xegaEvalGene}.
        \item \code{$evalFail}:  FALSE. Set by the error handler(s)
                                 in package \code{xegaEvalGene} 
                                 in the case of failure.
        \item \code{$fit}:       Fitness.
        \item \code{$gene1}:     A vector of reals.
        }
}
\description{
\code{xegaDfInitGene} generates a random real-coded gene 
             with a given length.
}
\details{
In the real-coded representation of 
         package \code{xegaDf}, \emph{gene} is a list with 
         \enumerate{
         \item \code{$evaluated} Boolean: TRUE if the fitness is known.
         \item \code{$fit}       The fitness of the genotype of 
                                 \code{$gene1}         
         \item \code{$gene1}     a real vector (the genetopye).
         }

         This representation simplifies the implementation 
         af various optimizations
         and generalizations.
}
\examples{
xegaDfInitGene(lFxegaDfGene)

}
\references{
Price, Kenneth V., Storn, Rainer M. and Lampinen, Jouni A. (2005)
The Differential Evolution Algorithm (Chapter 2), pp. 37-134. 
In: Differential Evolution. A Practical Approach to Global Optimization.
Springer, Berlin.
<doi:10.1007/3-540-31306-0>
}
\concept{Intialization}
